Report versioned lib.so in cc --print-file-name given short name
E.g. `cc --print-file-name libc.so` reports /usr/lib/libc.so.96.1

This is a complement of the major.minor finding logic in DriverUtils.
`ld -lc -L/usr/lib` currently find the libraries with this logic.

To make things more obviously related the code was extracted into a
function which was copied over verbatim.

fine with mortimer@
ok patrick@