workaround link training problems on s3 resume with alder lake
The firmware from multiple lenovo alder lake machines claim there are
two eDP ports in the video bios table (VBT).

Which results in link training problems on resume from S3 with the
screen continually turning on and off.  Forcing port B to not be
recognised as eDP works around this.  I believe port A is the normal eDP
port.

https://gitlab.freedesktop.org/drm/intel/-/issues/5531
https://gitlab.freedesktop.org/drm/intel/-/issues/4950