remove the "tbl" suffix for a few man pages
Over a decade ago, the build infrastructure had special logic to process
man pages that ended with the suffix "tbl".

This infrastructure is long gone and the special naming for these man pages
is no longer needed.

Revert the naming of these man pages for consistency with all other man
pages in the tree. As a bonus, we remove a few lines from some of the
Makefiles making them simpler.

ok jmc@, and no objection from schwarze@