Unifdef MONT_WORD.
It does not make sense to use code that is slower, currently broken and
prevents the use of assembly Montgomery implementations.

This is the result of `unifdef -m -DMONT_WORD`, followed by some manual
clean up and the removal of the Ni bignum from BN_MONT_CTX (which was only
used in the non-MONT_WORD case).

ok miod@ tb@