Process up to four test vector files concurrently.
This avoids having a slow down when processing test vector files that only
have a single group. Note that the processing of test vector files is in
turn going to be rate limited by the number of concurrent test groups,
which means we do not need variable limits for vectors.

Reduces a Wycheproof regress run down to ~8 seconds on an Apple M1.