dont match quectel ec25 by vendor+product id
quectel seem to provide a sane and consistent set of functionality
built on top of the qualcomm qmi stuff. their linux drivers guide
says quectel modems provide a set of umsm usb interfaces and a
network interface that can be in qmi, ecm, or mbim mode.

if the modem is in mbim mode, it will present the mbim classes which
umb should be able to attach to without this explicit vendor+product
match (assuming umsm doesn't claim the interface first).

based on info in the Quectel LTE&5G Linux USB Driver User Guide V2.0

discussed with and tested by kevlo@
ok patrick@ sthen@ kevlo@