add rkusbphy(4), a driver for the usb2phy on rockchip SoCs.
the rkusbphy device has children nodes in the device tree which are
the actual phys. this driver mostly exists to wire those children
up as PHYs and turn the associated regulators on when a host
controller enables the PHYs. in the future it should enable clocks
and take ports out of suspend too.

i'm not enabling this yet because it's useless without some tweaks
in the usb host controller drivers.

ok kettenis@