clockintr: clockintr_cpu_init(): isolate struct clockintr_queue setup
We will always need to initialize the clockintr_queue struct.
However, the hardclock, schedclock, and statclock will not always be a
part of that struct.  Move the clockintr_establish() calls in
clockintr_cpu_init() away from the other initialization steps to make
the dependency relationship more obvious.  We need to intialize a
struct clockintr_queue before we can establish clock interrupts.